+++ Changelog.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The versioning scheme now adheres to '0.major.minor.patch' to correctly indicate impacts of new releases. (#54)
 - The 'communication' module is renamed to 'clients'.
 ### Deprecated
+- Flask-webservice related functionality will be moved to a dedicated plugin. (#53)
 ### Removed
 - communication.module() and httprequests/localserver.proxy() functions.
 ### Fixed